refactor: 폴더 구조 변경, modal 띄울 떄 boolean 으로 onoff
