+++ src/atoms/modal/atom.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const modalAtom = atom<boolean>({
+  key: 'modalAtom',
+  default: false,
+});
+
+export default modalAtom;
