+++ src/hooks/useModal.ts
@@ -1,25 +1,20 @@
-import modalListAtom from '@/atoms/modalList';
-import { Modal } from '@/atoms/modalList/atom';
+import modalAtom from '@/atoms/modal/atom';
 import { useRecoilState } from 'recoil';
 
 interface UseModal {
-  openModal: ({ key, props }: Modal) => void;
+  openModal: () => void;
   closeModal: () => void;
 }
 
 const useModal = (): UseModal => {
-  const [modalList, setModalList] = useRecoilState(modalListAtom);
+  const [modal, setModal] = useRecoilState(modalAtom);
 
-  const openModal = ({ key, props }: Modal) => {
-    const newModalList = [...modalList];
-    newModalList.push({ key, props });
-    setModalList(newModalList);
+  const openModal = () => {
+    setModal(true);
   };
 
   const closeModal = () => {
-    const newModalList = [...modalList];
-    newModalList.pop();
-    setModalList(newModalList);
+    setModal(false);
   };
 
   return { openModal, closeModal };
